+++ examples/icon-eth-notification/src/pages/Navigate.js
