+++ examples/icon-eth-notification/src/pages/SignUp.js
